ml-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
google-generativeai>=0.5.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pypdf>=4.1.0
faiss-cpu>=1.8.0
sentence-transformers>=2.6.0
scikit-learn>=1.4.0
reportlab>=4.1.0
pytesseract>=0.3.10
python-dotenv>=1.0.1
ultralytics
opencv-python
